large update to allow for a gamestate description language
major changes: split parser directory back up (haha didn't i just undo that?
but this time it's so that i can create two different languages: TLL (Transaction Listing Language) and GSDL (Gamestate Description Language)

GSDL is a directive-based language that facilitates a thorough description of a gamestate for Treasuror purposes.
this code includes an example file as well as a parser for GSDL. An emitter will come later.

it's not wired in yet, oof. but it does compile. i will hopefully make it work soon.