Add a new operator (<>) which we use when a trade takes place between two parties.
This commit also restructures the lexer slightly.
Before, it did some of the work of the parser. Now, it shouldn't.
Next goal will include hopefully normalizing some of the inconsistencies in how code is patterned.
Finally, we will get to the name and goal of this branch, that being a completely new way to input commands, involving parenthesis.